+++ lista.c
@@ -53,9 +53,9 @@ void destroiLista(lista *lst)
 /* Recebe um ponteiro para a lista, a palavra buscada, e um ponteiro para um ponteiro de
 um no vazio. Retorna o numero de comparacoes para a busca e passa o endereco da palavra
 encontrada (ou NULL), por meio de ponteiro, para o resultado, fora da funcao */
-int buscaLista(lista *lst, char elemento[], no_lista **resultado)
+long long int buscaLista(lista *lst, char elemento[], no_lista **resultado)
 {
-    int cmp = 0;                 // Contador de comparacoes
+    long long int cmp = 0;       // Contador de comparacoes
     no_lista *p = lst->primeiro; // Ponteiro auxiliar de no
 
     // Checa se a lista esta vazia
@@ -95,10 +95,10 @@ linha de onde ela veio. Se a palavra ja existir na lista, apenas adiciona a nova
 ocorrencia no no da palavra antiga, se nao, a insere de modo a manter a ordem alfabetica.
 Alem disso, padroniza a palavra para a sua versao em letras minusculas. Retorna o numero
 comparacoes realizadas na operacao */
-int insereLista(lista *lst, char elemento[], int n_linha)
+long long int insereLista(lista *lst, char elemento[], int n_linha)
 {
-    int cmp = 0;     // Contador de comparacoes realizadas em insereLista
-    int relacao = 0; // Para armazenar o valor de strcmp
+    long long int cmp = 0; // Contador de comparacoes realizadas em insereLista
+    int relacao = 0;       // Para armazenar o valor de strcmp
     // Ponteiros auxiliares
     no_lista *p = lst->primeiro;
     no_lista *anterior = NULL;
@@ -190,9 +190,9 @@ int insereLista(lista *lst, char elemento[], int n_linha)
 /* Recebe a variavel txt que armazena o texto e um ponteiro para um ponteiro vazio de
 lista ligada. Retorna o numero de comparacoes realizadas para a criacao do indice e
 aponta o ponteiro de ponteiro de lista recebido para o indice final */
-int paraLista(txt texto, lista **lst0)
+long long int paraLista(txt texto, lista **lst0)
 {
-    int cmp = 0;
+    long long int cmp = 0;
     char *token;               // Ponteiro que recebe o endereco dos tokens
     lista *lst1 = criaLista(); // Ponteiro de lista ligada que sera retornado
